Vision Workflow Optimizer Agent/requirements.txt
fastapi
uvicorn[standard]
google-genai
python-dotenv
pydantic
google-adk